+++ Dockerfile
@@ -1,3 +1,3 @@
 FROM adoptopenjdk/openjdk11:ubi
-ADD target/test-service-0.1.1-SNAPSHOT.jar /app.jar
+ADD target/demo-service-0.1.1-SNAPSHOT.jar /app.jar
 ENTRYPOINT ["java","-jar","/app.jar"]
