+++ src/main/java/com/wine/to/up/demo/service/exception/KafkaRuntimeException.java
@@ -1,0 +1,11 @@
+package com.wine.to.up.demo.service.exception;
+
+public class KafkaRuntimeException extends FormattedRuntimeException {
+    public KafkaRuntimeException(String messagePattern, Object... params) {
+        super(messagePattern, params);
+    }
+
+    public KafkaRuntimeException(String messagePattern, Throwable cause, Object... params) {
+        super(messagePattern, cause, params);
+    }
+}
