+++ src/main/java/com/wine/to/up/demo/service/kafka/KafkaMessageHandler.java
@@ -1,0 +1,6 @@
+package com.wine.to.up.demo.service.kafka;
+
+public interface KafkaMessageHandler<Message> {
+
+    void handle(Message message);
+}
