+++ src/main/java/com/wine/to/up/demo/service/kafka/NamedThreadFactory.java
@@ -1,0 +1,19 @@
+package com.wine.to.up.demo.service.kafka;
+
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.atomic.AtomicInteger;
+
+public class NamedThreadFactory implements ThreadFactory {
+    private final AtomicInteger idGenerator = new AtomicInteger(1);
+    private final String prefix;
+
+    public NamedThreadFactory(String prefix) {
+        this.prefix = prefix;
+    }
+
+    public Thread newThread(Runnable r) {
+        Thread thread = new Thread(r);
+        thread.setName(this.prefix + "-" + idGenerator.getAndIncrement());
+        return thread;
+    }
+}
