+++ src/main/java/com/wine/to/up/demo/service/repository/MessageRepository.java
@@ -1,0 +1,11 @@
+package com.wine.to.up.demo.service.repository;
+
+import com.wine.to.up.demo.service.domain.entity.Message;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface MessageRepository extends CrudRepository<Message, UUID> {
+}
