--- src/main/java/com/wine/to/up/service/ServiceApplication.java
@@ -1,17 +1,0 @@
-package com.wine.to.up.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@SpringBootApplication
-@ComponentScan("com.wine.to.up")
-@EnableSwagger2
-public class ServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ServiceApplication.class, args);
-    }
-
-}
