--- src/main/java/com/wine/to/up/service/configuration/EventSerializer.java
@@ -1,17 +1,0 @@
-package com.wine.to.up.service.configuration;
-
-import com.wine.to.up.api.message.KafkaServiceEventOuterClass.KafkaServiceEvent;
-import org.apache.kafka.common.serialization.Serializer;
-
-/**
- * Serializer for {@link KafkaServiceEvent}
- */
-public class EventSerializer implements Serializer<KafkaServiceEvent> {
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public byte[] serialize(String topic, KafkaServiceEvent data) {
-        return data.toByteArray();
-    }
-}
