--- src/main/java/com/wine/to/up/service/configuration/GeneralPurposesBeansConfig.java
@@ -1,26 +1,0 @@
-package com.wine.to.up.service.configuration;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class GeneralPurposesBeansConfig {
-
-    /**
-     * Model mapper bean
-     */
-    @Bean
-    public ModelMapper getModelMapper() {
-        return new ModelMapper();
-    }
-
-    /**
-     * Object mapper bean
-     */
-    @Bean
-    public ObjectMapper getObjectMapper() {
-        return new ObjectMapper();
-    }
-}
