--- src/main/java/com/wine/to/up/service/kafka/KafkaMessageHandler.java
@@ -1,6 +1,0 @@
-package com.wine.to.up.service.kafka;
-
-public interface KafkaMessageHandler<Message> {
-
-    void handle(Message message);
-}
