--- src/main/java/com/wine/to/up/service/repository/MessageRepository.java
@@ -1,11 +1,0 @@
-package com.wine.to.up.service.repository;
-
-import com.wine.to.up.service.domain.entity.Message;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface MessageRepository extends CrudRepository<Message, UUID> {
-}
