+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=service-app
-spring.kafka.consumer.group-id=wine.test-service
+spring.kafka.consumer.group-id=wine.demo-service
 
 management.endpoints.web.exposure.include=prometheus
 management.endpoint.prometheus.enabled=true
